:recycle: refactor LanguageChanger localInfo handling by LOCALE
