--- src/components/LanguageChanger.tsx
@@ -1,41 +1,0 @@
-'use client';
-
-import { useTranslation } from 'react-i18next';
-import { usePathname, useRouter } from 'next/navigation';
-import { i18nConfig } from '@/i18nConfig';
-import { ChangeEvent } from 'react';
-
-export default function LanguageChanger() {
-  const { i18n } = useTranslation();
-  const currentLocale = i18n.language;
-  const router = useRouter();
-  const currentPathname = usePathname();
-
-  const handleChange = (e: ChangeEvent<HTMLSelectElement>) => {
-    const newLocale = e.target.value;
-
-    const days = 30;
-    const date = new Date();
-    date.setTime(date.getTime() + days * 24 * 60 * 60 * 1000);
-    const expires = date.toUTCString();
-
-    document.cookie = `NEXT_LOCALE=${newLocale};expires=${expires};path=/`;
-
-    if (currentLocale === i18nConfig.defaultLocale) {
-      router.push('/' + newLocale + currentPathname);
-    } else {
-      router.push(
-        currentPathname.replace(`/${currentLocale}`, `/${newLocale}`)
-      );
-    }
-
-    router.refresh();
-  };
-
-  return (
-    <select onChange={handleChange} value={currentLocale}>
-      <option value="ko">한국어</option>
-      <option value="en">English</option>
-    </select>
-  );
-}
